requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.28.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
